+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# [1.7.0](https://github.com/CiaranWoodward/LD56/compare/v1.6.0...v1.7.0) (2024-10-06)
+
+
+### Features
+
+* implemented hud ([80d2692](https://github.com/CiaranWoodward/LD56/commit/80d2692dd4b430f4e226663ea1ce5c7fc0fdc9e5))
+
+# [1.6.0](https://github.com/CiaranWoodward/LD56/compare/v1.5.0...v1.6.0) (2024-10-06)
+
+
+### Bug Fixes
+
+* crush speed if you faceplant the quarter-pipe wall ([8a0f92e](https://github.com/CiaranWoodward/LD56/commit/8a0f92efe132214a04a3c4bd16c2b2b897be6c80))
+
+
+### Features
+
+* Don't leap off quarter pipe if you don't have enough speed ([4881d29](https://github.com/CiaranWoodward/LD56/commit/4881d295410ff618716620cb3434541449427f7f))
+
 # [1.5.0](https://github.com/CiaranWoodward/LD56/compare/v1.4.1...v1.5.0) (2024-10-06)